requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.1.0
babel==2.17.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
dotenv==0.9.9
et_xmlfile==2.0.0
fastapi==0.115.9
fluent.runtime==0.4.0
fluent.syntax==0.19.0
frozenlist==1.5.0
future==1.0.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.5
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
netaddr==1.3.0
openpyxl==3.1.5
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.3.0
pycups==2.0.4
pydantic==2.10.6
pydantic_core==2.27.2
PyPDF2==3.0.1
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
yarl==1.18.3
yookassa==3.5.0
